Adjust console chart schema
Allow className and targetPort to be null.
Adjust generated test cases for `fsGroupChangePolicy` to fixed `Always` value.
Adjust deployment, ingress and service as new type is introduced that extends jsonschema.

+++ charts/console/service.go
@@ -28,7 +28,7 @@ func Service(dot *helmette.Dot) *corev1.Service {
 	}
 
 	if values.Service.TargetPort != nil {
-		port.TargetPort = intstr.FromInt32(*values.Service.TargetPort)
+		port.TargetPort = intstr.FromInt32(int32(*values.Service.TargetPort))
 	}
 
 	if helmette.Contains("NodePort", string(values.Service.Type)) && values.Service.NodePort != nil {
